Preserve 64-bit native Media I/O lease ABI

+++ src/Media/rtaime.Media/NativeMediaIoAdapter.cs
@@ -464,7 +464,7 @@ public sealed class NativeMediaIoProviderAdapter : IMediaIoProviderAdapter
 	{
 		public NativeIdentity LeaseId;
 		public ulong OpaqueHandle;
-		public uint ByteLength;
+		public ulong ByteLength;
 		public uint RowBytes;
 		public uint TransferMode;
 		public uint StorageDomain;
